introduce enum to objects.py
